scope oc search to only directories on path
Fixes https://github.com/openshift/jenkins-client-plugin/issues/234
